refactor: decompose email builtin package (MR B2)
- Split email/__init__.py (601 lines) into facade (302), models (99), executor (298)
- Split email/gmail.py (775 lines) into gmail/ package: facade (221), compose (79),
  fetch (398), labels (68), attachments (67), base (70)
- Extract EmailToolExecutor with all 8 async tool bodies
- Extract GmailComposer, GmailFetcher, GmailLabelManager, GmailAttachmentHandler
- Move tests to tests/builtins/email/ and tests/builtins/email/gmail/ with conftest.py
- All 2,913 tests passing, ruff clean, no new mypy errors
- Backwards compatible imports preserved
