+++ tests/builtins/email/__init__.py
@@ -1,0 +1,1 @@
+# Tests for email builtin tools
