+++ tests/builtins/email/gmail/__init__.py
@@ -1,0 +1,1 @@
+# Tests for Gmail provider
